Fixed PC-relative move instructions
PC relative move instructions now properly have a 4-byte offset--referencing
the PC of the next instruction instead of the PC of the current instruction.

Also adjusted the CIS combine-tester to prohibit moving any MOV instruction
that references the PC into the first of the two instructions.

Finally, strengthened the OP_ADDRESS decoding so as not to produce a segfault
when the address isn't given