Fixed an address to LDI conversion when the segment address is unknown
The problem had to do with a LDI label,Rx instruction.  If the label was known
to be within the current segment, an LDI instruction was issued--even before
it was known that the current segment would fit within the 23-bits of an
LDI instruction.

The solution is in two parts.  First, the LDI-MOV conversion was adjusted to
handle 15 bits of offset instead of 13, since PC offsets assume the bottom
two bits are zero anyway.  Second, for those instructions where a label is
used, and where it doesn't fit into a MOV, a full 32-bit LDI instruction
was chosen to be used instead.