Simplified the integer divide unit
I heavily simplified the big ugly state transition always block, removing
everything out of it that didn't need to be in it.  I then adjusted the
o_quotient value so that it wasn't written by index, but rather by shifting
new values from the end.  As for the transitions on (i_wr), I removed most
of the i_wr requirements--since many things can be always set--even if
nothing is changing within the core.  While this may increase the power
required, it also reduces the logic.

The good news is that the number of LUTs needed for the divide dropped by about
100 LUTs or so.  If for no other reason, because of this difference and this
difference only these changes are being kept and (more than that) checked in.