Modified the single-fetch prefetch unit to have a consistent calling convetion
This change is actually much more than that, as much of the internals to the
prefetch module were also simplified at the same time.  Among the
simplifications, reset was applied to the valid, cyc, and stb signals, but
no longer to the address line.  Further, to support compressed instructions,
the single-prefetch was made to be pipelined, to increment its address on its
own, and to fetch an instruction any time the last one was accepted.