Non-pipelined mode is now still pipelined pre-fetch through decode
This was really necessary to support CIS instructions in the non-pipelined
version of the CPU.  Changes include 1) adding a new dblfetch prefetch
module that will fetch two instructions at once in a pipelined fashion,
2) Moving the decode logic into the decoder, 3) making all the prefetch
modules have the same semantics.  Some other pipeline changes involve
making sure the op section clocks only on op_ce, since the decode data
might change when you aren't looking otherwise.  The NO_USERMODE option
was simplified to rely all the more on the optimizer, and so much of the
NO_USERMODE code that could be derived by an optimizer was removed.