Fixed GCC LW:DI (R0),R0 bug
This fixes a GCC bug that caused loading a DI register from address R0 into
R0,R1 from producing the code:
	LW  (R0),R0
	LW 4(R0),R1
Our solution was to change that into,
	LW 4(R0),R1
	LW  (R0),R0
While this defeats the pipelined optimizations, it also still works.  Ideally,
we'd want to load into another register, but ... that fix will be the longterm
one.