Fixed the ZipCPU's problems with running the testbench
Test bench now works for both stepping, as well as running with single fetch
(no pipelining), double fetch, and regular cache'd prefetch.

Fixed lots of bugs all around.

The instruction decoder now produces a strobe signal for early branching,
in addition to the decoded wire associated with the instruction.  The
strobe is for the prefetch which will act on it immediately, the other
for the logic which may be busy for a while.

Aux busdelay: dropped an excess wire, added logic to handle the case where
the buffer is empty and the output is stalled.

Fixed an ugly bug in dblfetch, causing it to output an old out of date
instruction together with the valid bit high.

The test bench has been renamed zipcpu_tb.cpp, which produces both
zipsys_tb and zipbones_tb.

The test program, simtest.s, was also modified so as to use the NOOP
version of the simulation instructions--the instructions which won't crash
the simulator on an illegal instruction.