Updated Simtest
Fixed trap -1 statements, which weren't trapping but rather putting the CPU
to sleep.  A future modification to the binutils will make certain the
GIE bit is cleared whenever a trap command is issued--this wasn't.

Once the trap was fixed, it then became possible to see that there were
other bugs that hadn't been properly fixed--primary among these was a
misalignment bug.