Exported wb counts, stalled pipemem at 13
Exported the formal wb counts, f_nreq, f_nack, and f_outstanding, if FORMAL
is defined.  This allows these components to be connected via assertions,
allowing induction to get farther.

Also, stalled pipemem any time 13 items are in the pipe.  While I had
originally thought this would never happen, it could actually happen
in a massive memset program that fits within the i-cache.  This will
slow the CPU down in this case, but the difference may not be relevant.