ZipAXI + ZipCore can now gate clock, AXIDCACHE fix
AXIDCACHE now only ever issues 32-bit writes.  This became an issue with
SWAP_WSTRB.  Fixing this was intrusive.  ZipAXI needed a change as well.

IDecode now has an (unused) NULL branching option.  When/if set, unconditional
branches to the next instruction--skipping none--will quietly be turned into
NOOPs.  While this can greatly improve performance, it will only do so if the
compiler is broken.  The compiler has now been fixed, so this will only ever
likely be an unused option.

AXILFETCH is now instruction word size independent.  This was necessary to
support another program.